+++ packages/lumx-react/src/components/button/ButtonGroup.tsx
@@ -1,32 +1,14 @@
-import { GenericProps } from '@lumx/react/utils/type';
-import type { LumxClassName } from '@lumx/core/js/types';
-import { classNames } from '@lumx/core/js/utils';
 import { forwardRef } from '@lumx/react/utils/react/forwardRef';
 
-/**
- * Defines the props of the component
- */
-export interface ButtonGroupProps extends GenericProps {
-    /**
-     * Children
-     */
-    children?: React.ReactNode;
-}
+import {
+    ButtonGroup as UI,
+    ButtonGroupProps,
+    CLASSNAME,
+    COMPONENT_NAME,
+    DEFAULT_PROPS,
+} from '@lumx/core/js/components/Button/ButtonGroup';
 
-/**
- * Component display name.
- */
-const COMPONENT_NAME = 'ButtonGroup';
-
-/**
- * Component default class name and class prefix.
- */
-const CLASSNAME: LumxClassName<typeof COMPONENT_NAME> = 'lumx-button-group';
-
-/**
- * Component default props.
- */
-const DEFAULT_PROPS: Partial<ButtonGroupProps> = {};
+export type { ButtonGroupProps };
 
 /**
  * ButtonGroup component.
@@ -36,14 +18,9 @@ const DEFAULT_PROPS: Partial<ButtonGroupProps> = {};
  * @return React element.
  */
 export const ButtonGroup = forwardRef<ButtonGroupProps, HTMLDivElement>((props, ref) => {
-    const { children, className, ...forwardedProps } = props;
-
-    return (
-        <div ref={ref} {...forwardedProps} className={classNames.join(className, CLASSNAME)}>
-            {children}
-        </div>
-    );
+    return UI({ ref, ...props });
 });
+
 ButtonGroup.displayName = COMPONENT_NAME;
 ButtonGroup.className = CLASSNAME;
 ButtonGroup.defaultProps = DEFAULT_PROPS;